requirements.txt
certifi==2020.6.20
cffi==1.14.0
cryptography==3.0
decorator==4.4.2
pycparser==2.20
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
emoji==0.5.4
clarifai==2.6.2